Move inode related functions to inode.rs
